backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.24.0
supabase>=1.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.1.0
